refactor: 프로젝트 폴더 구조 정리 및 코드 리팩토링
