--- src/main/java/com/example/api_server/dto/request/SignUpReqDto.java
@@ -1,19 +1,0 @@
-package com.example.api_server.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SignUpReqDto {
-    @NotBlank(message = "아이디는 필수 입력값입니다.")
-    private String username;
-
-    @NotBlank(message = "비밀번호는 필수 입력값입니다.")
-    private String password;
-}
