--- src/main/java/com/example/api_server/entity/RoomParticipant.java
@@ -1,54 +1,0 @@
-package com.example.api_server.entity;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class RoomParticipant {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "room_participant_id")
-    private Long roomParticipantId;
-
-    @Column(name = "room_id", nullable = false)
-    private Long roomId;
-
-    @Column(name = "user_id", nullable = false)
-    private Long userId;
-
-    @CreationTimestamp
-    @Column(name = "joined_at", nullable = false, updatable = false)
-    private LocalDateTime joinedAt = LocalDateTime.now(); // 방 참여 시간
-
-    @UpdateTimestamp
-    @Column(name = "updated_at", nullable = false)
-    private LocalDateTime updatedAt; // 마지막 업데이트 시간
-
-    // 생성자
-    public RoomParticipant(Long roomId, Long userId) {
-        this.roomId = roomId;
-        this.userId = userId;
-    }
-}
