--- src/main/java/com/example/api_server/global/config/SwaggerConfig.java
@@ -1,24 +1,0 @@
-package com.example.api_server.global.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.Components;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .components(new Components())
-                .info(apiInfo());
-    }
-
-    private Info apiInfo() {
-        return new Info()
-                .title("Image-Editor Rest APi") // API의 제목
-                .description("Image-Editor와 통신을 위한 Rest API") // API에 대한 설명
-                .version("v2"); // API의 버전
-    }
-}
