--- src/main/java/com/example/api_server/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.example.api_server.repository;
-
-import com.example.api_server.entity.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-    boolean existsByUsername(String accountName);
-
-}
