+++ src/main/java/com/example/api_server/user/dto/request/SignUpReqDto.java
@@ -1,0 +1,19 @@
+package com.example.api_server.user.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class SignUpReqDto {
+    @NotBlank(message = "아이디는 필수 입력값입니다.")
+    private String username;
+
+    @NotBlank(message = "비밀번호는 필수 입력값입니다.")
+    private String password;
+}
